# Use vctrs for flattening into atomic vectors