--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -21,6 +21,7 @@ BugReports: https://github.com/tidyverse/purrr/issues
 Depends: 
     R (>= 3.2)
 Imports: 
+    ellipsis (>= 0.3.1),
     lifecycle (>= 0.2.0),
     magrittr (>= 1.5.0),
     rlang (>= 0.4.7),