--- a/R/utils.R
+++ b/R/utils.R
@@ -326,3 +326,8 @@ quo_is_same_env <- function(x, env) {
   quo_env <- quo_get_env(x)
   is_reference(quo_env, env) || is_reference(quo_env, empty_env())
 }
+
+unstructure <- function(x) {
+  attributes(x) <- NULL
+  x
+}