--- a/src/init.c
+++ b/src/init.c
@@ -29,7 +29,6 @@ static const R_CallMethodDef CallEntries[] = {
   {"map2_impl",             (DL_FUNC) &map2_impl,      5},
   {"pmap_impl",             (DL_FUNC) &pmap_impl,      4},
   {"transpose_impl",        (DL_FUNC) &transpose_impl, 2},
-  {"vflatten_impl",         (DL_FUNC) &vflatten_impl,  2},
   {"purrr_eval",            (DL_FUNC) &Rf_eval,        2},
   {NULL, NULL, 0}
 };