--- a/tests/testthat/helper-map.R
+++ b/tests/testthat/helper-map.R
@@ -1,5 +1,10 @@
 
-named <- function(x) set_names(x, chr())
+named <- function(x) {
+  if (is_null(names(x))) {
+    names(x) <- names2(x)
+  }
+  x
+}
 
 # Until we can reexport from rlang
 vars <- function(...) rlang::quos(...)