… src/flatten.c