@DavisVaughan 2020-08-07
It might be worth revisiting this if `vec_unchop()` has gotten faster 
https://github.com/tidyverse/dplyr/blob/dc3e0e6ccc54b049ecd89ca83455cdb101e6583d/R/join-rows.R#L54