@lionel- 2020-08-07
which makes total sense since the likely bottleneck is the common type determination.