@DavisVaughan 2020-08-07
I think most of it comes from the casting and the assignment. I feel like we could probably make the assignment faster for at least the atomic cases, but that's just a guess.

<img width="616" alt="Screen Shot 2020-08-07 at 4 32 59 PM" src="https://user-images.githubusercontent.com/19150088/89686105-b384b180-d8cb-11ea-9fbd-d0248c471a74.png">