@lionel- 2020-08-10
I now wonder if `"{inner}"` should be our general default in vec-c / list-unchop:

- No transformation of data.
- Outer names are ignored (lost data), but the more specific (and thus probably more relevant) inner names are kept.
- No errors which might be annoying during data analysis.