@lionel- 2020-08-10
Actually `"{inner}"` doesn't work as a default because `inner` might be numeric if the elements are unnamed.

I fixed it but there is currently no way to zap names from a name-spec so this is a behaviour change:

```r
# CRAN
names(purrr::flatten_int(list(x = c(1L, 2L), 3L)))
NULL

# This PR
names(flatten_int(list(x = c(1L, 2L), 3L)))
#> [1] "" "" ""
```

Feature request at https://github.com/r-lib/vctrs/issues/1215.