@hadley 2022-08-26
We should consider renaming to `list_flatten()` (with a `.ptype` argument) and deprecating `flatten_*`, if this will affect much existing code.