@hadley 2022-09-04
Superseded by #912.